Prototype of splitting provisioning
Based on code from vendor/google/apps/NfcProvision.

It uses a pre-specified action and a package name specified in the bump to
launch the EDU specific pieces.  It launches that activity and waits for a
broadcast back in both the pre-MDM and post-MDM steps.  Still needs the
security check to make sure that the activity being launched is part of a
privileged app.

The MDM is currently specified by package and admin receiver in the nfc bump.
This needs to be expanded to allow a download location rather than a package
to enable and use.

Change-Id: Iaabe8505c5e33c33ddec3122d2a17b07a6b243fc