Retain only requiredForAllUsers apps when creating a managed profile.
When a managed profile is created, system apps are installed for the new user by default.
This change removes all system apps except for those marked as requiredForAllUsers in their manifest.
Additionally, a few hard-coded system apps are retained as a temporary measure to enable development & testing.
These will be removed in a later CL.

Change-Id: Ic24b2d5ca9d119c2ccfd7ef34ff7d298594c9434