Remove string that is not user visible.
Avoids additional translation work or having a string that is untranslated in case we ever decide to show the app name anywhere. Only name this may show up at the moment is under running apps in settings and it is consistent with other apps not to have a name there (package name shown instead).

Change-Id: I36120858138c3a836ea5def6622b5c42629c4810