Allow to specify the admin component for provisioning.
An app may have several device admins.
For this case, allow the caller to specify which one will become
an active device admin, and the profile owner (in the profile owner flow).

We still support the previous way which was to specify only the package.
In this case, we continue to infer the admin from the manifest.

BUG:19348295
Change-Id: I91a74afbad809e838fb6807e182f66f189b7b356