Persist parameters during final stage of Device Owner provisioning.
During the final stage of DO provisioning, a Device Initialisation Agent (DIA)
will run and in the unlikey event of device accidentally rebooting, we should
make sure the flow continues after reboot. DIA of course needs to handle that,
but this also means we need to persist the ProvisioningParams so when DIA
finishes, we can still continue the last step of firing
ACTION_PROFILE_PROVISIONING_COMPLETE to MDM, in HomeReceiverActivity.

Bug: 19806407
Change-Id: Ia5d43626282166e71f479b24aebb5865a6a558c1