Wait for proxy to start.
If Bluetooth fails to start, an ANR may surface. With this change, the
BluetoothConnectTask will wait for the BluetoothConnectionService to send a
notification indicating that the proxy has been set up. This allows the task to
block until the network is available and allows the proxy to be set up on a
background thread.

The BluetoothConnectTask now waits for for the proxy to be set up before the
timeout is triggered. The timeout value has been increased to compensate.

Bug: 20455939
Change-Id: Ie2928cbfcf114ae5331ca6e183f4ee34bec4d2f6