Refactore the IntentStore.
The previous code required to specify all the extras that may be
present in the intent we wanted to save.
This was error prone, because if we wanted ManagedProvisioning to
support a new extra, and if we forgot to specify it had to be stored,
this extra would be forgotten after encryption.
Instead, just loop over the intent extras, and save each one.

The previous code also specified a different intent store for the
profile owner and device owner cases.
Instead, just use the action of the intent stored to differentiate
between the profile and device owner cases.

BUG: 25836548
Change-Id: I2ed34ac1126a2e0a5cb0234eea475575a5cd77eb