DO NOT MERGE: Only select Wi-fi when needed
When doing an NFC provisioning where the DPC is already installed on the
device, the user should not have to select a Wi-Fi. This situation can
happen when an NFC provisioning happens without providing
EXTRA_PROVISIONING_DEVICE_ADMIN_PACKAGE_DOWNLOAD_LOCATION.

Test: Use NfcProvisioning app to initiate an NFC provisioning without
specifying an SSID and without setting a download location.

Bug:32629443
Change-Id: I0c689bacb3f980daaf9bd3e7ead85ab8e169c6c7