[1] Avoid jank between MP finishing and DPC launching
For PO provisioning we send an ordered broadcast, finish the
provisioning activity and wait for the broadcast to be received by the
DPC. This creates HOME screen to be launched for 2 sec.

With this change we have introduced a new activity intent action which
is launched when provisioning is successfull. The DPCs can resolve the
action and start their post provisioning tasks.

Test: runtest managed-provisioning

Bug: 30956199
Change-Id: Iaebd3697b2d461e36781af0421bdcfd6053b6686