Tidy up PreBootListener a little
Trying to tidy the listener up a little and adding test coverage for it.
Fixing a couple of bugs on the way. There's still a problem here, as
we never wait for task execution to complete and there is no guarantee
that our process won't get killed in the middle of it.

Bug: 27353778
Bug: 32079350
Test: runtest managed-provisioning
Change-Id: I8f39c0291bb58110dce19d924d59e29cb6de313b