To make testing trusted source provisioning possible
To achieve this, MP now allows current device owner to replace its APK that
is test-only.

To make sure it is safe, we only allow test-only package install if
the same package is the current device owner. The justification are
as follows:
1. When we reach there, package to be installed is already a test-only
   device-owner.
2. PKMS will make sure the replaced APK is signed with the same
   signature.

With both 1 and 2, we can ensure that MP is only replacing the APK of
the current test-only DO. And thanks to the fact that DPMS persists the
testOnlyAdmin flag, going from test-only <--> non-test-only should has no
effect.

Test: runtest managed-provisioning

Bug: 62563858

Change-Id: Ie8af7020aad452ede3c6eb44e586a153bc721112