Fix blank window is shown briefly when MP is cold started
We should use the windowDisablePreview to turn off the initial
blank screen that the system process draws when launching the app.
We did have this flag previously by using SuwThemeGlif.Light indirectly,
but it was removed  in ag/2401821.

Note that SetupWizard is also setting SuwThemeGlif.Light as the default
theme in AndroidManifest and have it overrided in runtime.

Fix: 63810809
Test: Force stop managed provisioning app, launch testdpc to carry out
provisioning. Make sure the blank window is not shown anymore.
Test: adb shell setprop setupwizard.theme glif_v2_light, run the above
      test again.

Change-Id: I4cb0738075265f19b0341d17bb45ab8660b1a009