Fix GLIF theme is not applied
Depend on the device, the GLIF theme could be v1, 2, 3 and we set the theme
in runtime to apply the correct theme by calling setTheme.

However, there was a bug and we override the theme that we just set
for tinting the "swiper". We should not set the activity theme
for just the color of swiper, we should just get the swiper drawable
with the correct theme.

FIXES: 70621340

Test: atest ManagedProvisioningTests
Test: Try setting customized color, observe the swiper is tinted properly.

Change-Id: I2eecb9c6896c30fadb0a55b4c4686b3c0ed4eb67