Lookup device admin component in specified user
The previous fix introduced an issue, NonRequiredAppsLogic is run
in OTA as well and the infer doesn't work if it is a work profile
and the DPC is no longer installed in user 0.

Two changes:
1. findDeviceAdmin now takes user id as the parameter, we should
   explicitly specify which user we are looking for admin package.

2. Introduce inferDeviceAdminComponentName, so it is more
consistent to inferDeviceAdminPackageName, and we only do infer
if component name is missing in the param. We should only call
Utils.findDeviceAdmin directly if findDeviceAdmin if it is for
verification purpose.

Test: Try DO sync auth
Test: Try creating work profile
Test: Try OTA in work profile case, while DPC is not installed in user 0
Test: atest ManagedProvisioningTests
Test: afw-test-tradefed-ci run cts --module AeTestNfcProvisioningTestCases -l DEBUG

FIXES: 74107215

Change-Id: I601e391f6db710a1f59d33d7c2cc554d9a835fde