Update patch set 1
Patch Set 1:

Assume you want me to fix so 1) It's using only spaces as whitespace and 2) Edit/remove the comment about Google at the end of the line?  

Please clarify if I miss-understood.

If you are asking to substitute AOSP packages instead of Google packages, I don't think this is possible.  Initializing an Android 'Work Profile' depends on Google Play Services or it fails.  As far as I know there is no equivalent AOSP for Google Play Services.

The above code was given (in Android documentation website) as a working minimum set of dependencies as used in a Nexus device.

Patch-set: 1