Add Robolectric tests for ConnectMobileNetworkTask.
This class in ManagedProvisioning was previously not tested.

This is the first Robolectric test in ManagedProvisioning, aside from
the proof-of-concept that bfranz@ created in WifiInfoRoboTest, which was
copied from an instrumentation test. This achieves the same goal as an
instrumentation test, but is far quicker to run, has far fewer
dependencies, and doesn't run on a device.

Also improve the Javadoc of AbstractProvisioningTask and ConnectMobileNetworkTask.

Bug: 79513466
Test: From /packages/apps/ManagedProvisioning, run "atest tests/robotests/src/com/android/managedprovisioning/task/ConnectMobileNetworkTaskRoboTest.java"
Change-Id: I8b5edc37fec1bf45537fce4ea4f267e6ebf89eb5