Use a service prevent MP process from dying while interacting with DPC.
Prior this CL, when we launched the DPC as part of the
ManagedProvisioning finalization steps, the process got killed as it
was considered inactive. We fix that by launching the DPC and waiting
for it to receive the PROFILE_PROVISIONING_COMPLETE broadcast. We
kill the service after that.

Bug: 111333517
Bug: 110245352
Test: gts-tradefed run commandAndExit gts-dev --module
GtsInstallPackagesWhitelistDeviceTestCases --t
com.google.android.installpackageswhitelist.gts.GtsInstallPackagesWhitelistDeviceTest#testInstallerPackagesAgainstWhitelist

Change-Id: I54c4b60fe7a35fb6438ecfe9faa684dc379accea