Accommodate DPM API change for Device ID access method
The method to set Device ID access for a specific user was changed in
the DPM to operate on the current user.

To accommodate, get the DPM of the managed profile user (from a context
associated with that user) and set Device ID access on that instance.

Bug: 126700461
Test: Manual, installed TestDPC on a device flashed with that build and
made sure TestDPC can generate keys with Device ID attestation.

Change-Id: I5d76f73c326d3158d255cb96923464f1a430a8e3