Check if callback is still desired from the main thread.
Inside the main thread, we have to first verify the callback is still
present on the callbacks list. This is because when a config change
happens (e.g. a different locale was specified), ProvisioningActivity
is recreated and the old ProvisioningActivity instance posts the
callback on the UI thread. That callback then gets processed after
the old instance's onStop() is called, leading to bad state.

Test: manual
Fixes: 131719633
Change-Id: I445cb84fed67464c718b4b3025e5b03d20ba91ff