Let DevicePolicyManager know when the device is provisioned.
Previously, we were relying on the OobConfig app to do this. Since
DEVICE_PROVISIONED was being set BEFORE letting DevicePolicyManager
know, this was causing unusual behaviours, particularly with ZeroTouch.

This currently isn't unit testable. I tried both Robolectric (we're
currently getting an error with theming that we haven't resolved yet)
and instrumentation (admin-integrated flow has not been unit-tested,
and it fails when there isn't a DPC to launch).

Fixes: 131749749
Fixes: 130443862
Test: manual
Change-Id: I6563ba621cf16d3451dc191fe5458ba9fbd4e554