Refactor to update the provisioning state during gellerization.
During the gellerization, the user had already been through the unmanaged provisioning flow once.
This means the user_setup_complete will be already set, yet the provisioning should still be valid,
as opposed to other cases where user_setup_complete will be set.

This CL makes possible to update the provisioning state during the gellerization to signal that the provision is
indeed valid, and should proceed. Otherwise the flow will be interrupted and the device will be
left unmanaged.

Bug: 136148966, 135048328
Test: runtest managed-provisioning
Change-Id: I8b57619c25527af854669e0ea678c039bffa1807