Restrict removal of managed profile created during SUW
When the managed profile is created as part of the admin-integrated
flow, during the initial device set-up, and is part of the
organization-owned provisioning flow, then restrict its removal.

The restriction is applied by setting the DISALLOW_REMOVE_MANAGED_PROFILE
directly on the UserManager. This way, the restriction would be a part
of the base restrictions set by the system (as opposed to the DPC) and
so no UserManager changes are needed for it to apply.

It is not possible to set the restriction on the active admin of the
primary profile because it does not have one - the only DPC allowed in
this mode is that of the managed profile.

With this change, there's no way to remove the managed profile but to
wipe the entire device.
A new DPM API will soon be added to relinquish control of a device which
will remove the restriction and immediately wipe the profile.

A new test was added that verifies both application of the user
restriction as well as granting access to Device IDs (to the managed
profile DPC).

Bug: 138709470
Test: atest ManagedProvisioningTests:com.android.managedprovisioning.finalization.FinalizationControllerTest
Change-Id: I5508fc4269c050637adb8426599424a48a54d0cc