Revert TimeZoneDetector / TimeDetector changes
This reverts commit fddd9df6a003f55e71d0c7b232add0b13be22f76.
and commit e0eee6d208485734ea52c4cada8ebf0498fb49c0.

It's not clear that managed provisioning should be treated as an end
user setting the time / time zone. While we work it out, revert the
changes that have been landed already.

Bug: 140712361
Bug: 146437903
Test: treehugger
Exempt-From-Owner-Approval: It's a revert and OWNERS are out
Change-Id: I6b34e419ca921ebbedb317a151318b1df722f38c