Cherry-pick: Remove usage of WifiConfiguration#setProxy()
This is a cherry-pick of ag/10204741 to AOSP to solve
conflicts. See ag/10204741 for comment history.

setProxy() can be replaced with methods on
IpConfiguration itself. There is no need to expose
this as a @SystemApi, and is being removed.

Bug: 146046526
Test: atest WifiConfigurationProviderTest
Change-Id: I4e1bff220af11e6ce9e6d8553ad9c3ca9db1c5f7
Merged-In: I4e1bff220af11e6ce9e6d8553ad9c3ca9db1c5f7