Add provisioning cross-profile consent screen
This is just the initial CL to create the UI so it can be placed into
the provisioning flow.

The 'next' or 'done' button currently does not set the app-op. Unit
tests are not covering the button yet.

Changes in CustomizationParamsTest are just following the existing style
pattern.

Strings are marked as non-translatable since they are not yet finalized.

Bug: 136249261
Bug: 140728653
Bug: 148769796
Test: atest --verbose com.android.managedprovisioning.provisioning.crossprofile.CrossProfileConsentActivityRoboTest
Change-Id: Ie3a23ea1e8f0d6ed7781c75f30da39dc4585ad77