Use SavedStateHandle for provisioning cross-profile consent screen
This is the new ViewModel feature that means you can avoid listening to
onSaveInstanceState and onRestoreInstanceState in the activity, then
piping them through to the ViewModel. Although ViewModels are preserved
after configuration changes, they can still be destroyed and recreated
when the activity is destroyed by the system. This allows the state to
be preserved in that case.

Also, prevent double-clicks on the button; not critical here but it's
useful to get this working correctly from the beginning.

Bug: 136249261
Bug: 140728653
Bug: 148769796
Test: atest --verbose com.android.managedprovisioning.provisioning.crossprofile.CrossProfileConsentActivityRoboTest
Change-Id: Ib0de1582547c4ab49b4135e30c542130348e97ad