Set cross-profile app-ops from provisioning
Make the button functional and test it. Set the app-ops according to the
toggle states and set a shared preference for future reference. Check
for unusual race conditions where items being shown in the UI are not
understood by the model, or when the button is pressed too soon. Finish
early if there are no items to show in the screen.

Also, enable to commented-out test for the icon as a parallel CL fixes
the shadows to make this work.

Bug: 136249261
Bug: 140728653
Fixes: 148769796
Test: atest --verbose com.android.managedprovisioning.provisioning.crossprofile.CrossProfileConsentActivityRoboTest
Change-Id: Ib3bd0062e1cbbb4d2d70d86fce45af745fdfe9b1