Do not remove work profile if restricted
If removal of the work profile is restricted via the
DISALLOW_REMOVE_MANAGED_PROFILE user restriction, do not offer the user
the option of removing an existing profile when they try to set up a new
one.

This is not a comprehensive fix.
The initiateProvisioning should call checkDevicePolicyPreconditions in
all circumstances, otherwise we might discover other flows through which
the restriction to remove an existing, managed profile can be
circumvented.

Bug: 160305301
Bug: 146048940
Test: Manual, use TestDPC on a device provisioned in "managed profile on company-owned evice" to try and provision a new profile.
Test: atest ManagedProvisioningTests
Change-Id: I5916ed5252ac6b724aa1e65f9be9ffc918eecf07