Remove "delete work profile" dialog
Managed Provisioning "helpfully" offered the user the option to delete
an existing work profile if a new work profile provisioning was started.

Remove this dialog as it could cause confusion on, for example,
company-owned devices where the profile cannot be removed.

Additionally, move provisioning preconditions check to an earlier
stage in the provisioning process to avoid potential device state
modification if provisioning is not possible.

Bug: 146048940
Test: atest ManagedProvisioningTests
Change-Id: I854a71c27485a9f75b432f079210d2f3a8ed3e87