Refactor code for easier admin-integrated flow maintainability
This refactoring is meant to improve readability and
maintainability for the admin-integrated flow logic.

This helps reduce the risk of bugs for the upcoming
changes where the admin-integrated flow will be used
for sync auth as well.

Notable changes:
- Get rid of provisioningMode entirely - back in
Android 10 it was added as a temporary "hack" to
make it easier to add the admin-integrated flow logic.
That increases complexity, because in addition to that
we also have the pre-existing provisioningAction which
is now sufficient for us.
- Add flowType to indicate if we're doing the admin-
integrated flow or legacy one. In the future, when
we only support the admin-integrated flow and deprecate
the legacy one, this will not be necesary.
- Centralise the checks for whether to perform the
admin-integrated flow. Prior to this change, there
were several checks spread across the code to determine
whether we should do the admin-integrated flow or not.
This has now been centralized to
shouldPerformAdminIntegratedFlow.
- Move code away from big classes to dedicated utils
classes like GetProvisioningModeUtils and ParserUtils

Bug: 167955145
Test: TBD
Change-Id: I19188132c171f7063baf1381fc176771528a331a