Changed use of Activity to Context
This is a pre-requisite for the changes to the silent provisioning flow
in ArcManagedProvisioning.

All the methods invoked on mActivity are actually methods of the Context
class. As in ArcManagedProvisioning we will invoke a method of
PreFinalizationController from a Service derived class, we can safely
change the type parameter to Context, which is a common ancestor of
Activity and Service.

Bug: b/155974504
Author: giovax@
Test: Tested manually with a mix of unmanaged, managed and Unicorn users
Change-Id: Ic57d0b2ccf0ab2b6dfb53efa33f47aada462cde6
(cherry picked from commit 1b8d17cb8f168ecef5b202c5d7f4e8e7df8d877b)