Add the isSetupFlow extra set to true for NFC provisioning
This is a temporary solution to allow ManagedProvisioning use
OEM-customized themes. Usually it's the setup wizard who
sets this extra when starting ManagedProvisioning, but NFC
provisioning is different as ManagedProvisioning is the one
who starts itself, rather than setup wizard.

This fix will not be needed anymore when NFC provisioning
is updated to use the admin-integrated flow.

Bug: 132105112
Test: manual
Change-Id: Ia596b8a7d72277a041d0b067bc0e3baeb3d01820