Provisioning extra for admin permission grant opt-out
Support the EXTRA_PROVISIONING_PERMISSION_GRANT_OPT_OUT provisioning
extra for letting an admin opt out of controlling grants over
sensors-related permissions, on fully-managed devices.

Take advantage of the newly-introduced
FullyManagedDeviceProvisioningParams so that a new setter is not
required.

NOTE: A follow-up change is required the
FullyManagedDeviceProvisioningParams are not yet filled from
ProvisioningParams.

Bug: 158735247
Test: atest ManagedProvisioningTests:com.android.managedprovisioning.parser.ExtrasProvisioningDataParserTest ManagedProvisioningTests:com.android.managedprovisioning.parser.PropertiesProvisioningDataParserTest ManagedProvisioningTests:com.android.managedprovisioning.model.ProvisioningParamsTest
Change-Id: Iab08aaba5c47f58bd58ce7e93f47391a87d1a014