Stop using deprecated broadcasts in NetworkMonitor
Broadcasts like CONNECTIVITY_ACTION or INET_CONDITION_ACTION have been
deprecated in favor of network callbacks for a while. Migrate current
usage to the callbacks, which are more reliable and flexible.

Current usage of CONNECTIVITY_ACTION and INET_CONDITION_ACTION in
NetworkMonitor is just to get notified when a default network is
connected (as per NetworkInfo.isConnected). This is equivalent to
receiving onAvailable on registerDefaultNetworkCallback.

This change just keeps existing behavior, however considering current
usage of NetworkMonitor, it should probably be changed to verify that a
given network transport (wifi, cellular) is available when requested,
instead of checking that any default network is available. This change
makes it easier to do the behavior change by using requestNetwork with
a given transport type.

Test: atest ManagedProvisioningTests
Bug: 182859030
Change-Id: Ia56680d7b3a967d29f02b98c70033d4d670324e6