Copy wizard manager extras to intents that launch the admin app
More specifically, we copy the extras to the
ACTION_GET_PROVISIONING_MODE and ACTION_ADMIN_POLICY_COMPLIANCE
intents. This way ManagedProvisioning will forward the
setup wizard-specified intent extras to the admin app, and
the admin app can use the extras to determine if it should
apply a different theme, for example.

Fixes: 180699877
Test: manual
Change-Id: I4848ee67aef675365acb9ced6b917d26bd636eda