Make TermsActivity terms list take full width in landscape
Previously, TermsActivity was using GlifRecyclerLayout which
split the screen in two halves in landscape mode. UX asked
to change that to make it so that the terms list takes up
the entire width of the screen.

This CL also fixes the paddings and margins of the items
to be aligned with the mocks.

Fixes: 184627856
Fixes: 184628645
Test: manually tested in portrait and landscape mode with both
AOSP and Pixel-specific builds
Test: manually tested with light and dark mode

Change-Id: I2e714582bc3df7b00aadb1741b322ff63b92ef2a