Use a callback instead of  listening for a broadcast when installing apps
Listening for a broadcast is not reliable enough and we get scenarios
where the broadcast is received before installation is actually
completed. This causes provisioning to fail, because it tries
to launch a DPC it thinks is installed but it isn't.

This CL also revamps the InstallPackageTask test to test its
actual behaviour rather than implementation details.

Fixes: 185897624
Test: manual
Test: atest InstallPackageTaskTest
Change-Id: Id9eafa4a67bc25cad92b44f215220da766b3e79f