Set correct background color for deferred setup
ManagedProvisioning is not started with the isSetupFlow
extra for deferred setup managed account provisioning,
so we temporarily rely on a flag that the setup wizard sets.

Also we move the logic that sets isSetupFlow for NFC
in the same place before super.onCreate for consistency
and to avoid potential similar bugs there.

Fixes: 191273126
Test: manually went through deferred setup (videos in bug)
Test: manually tested NFC provisioning
Change-Id: Ic65c5ab41b177a35f38b78d45d02ce7f0362ab15