Wait for validated network for mobile data
ConnectMobileNetworkTask uses NetworkMonitor to receive network
connected notifications, but only completes if the network is validated.
However since change r.android.com/1647648, the NetworkMonitor will
not send notifications on validated status change.

Allow NetworkMonitor to wait for a network to be validated to send the
callback, and use that option in ConnectMobileNetworkTask.

Bug: 195499714
Test: atest ManagedProvisioningTests
Change-Id: I144517eec1ee985d1501b4e3fe7314e7d0c86ab3