Use the managed user to check for the existence of POLICY_COMPLIANCE
We previously used the system user to check whether the DPC has
the POLICY_COMPLIANCE handler. However, DPCs can disable themselves
in the personal profile. In such cases, the personal profile
POLICY_COMPLIANCE handler will not exist, making ManagedProvisioning
think that the DPC does not support the admin-integrated flow. The
POLICY_COMPLIANCE handler itself is launched in the work profile for
work profile provisioning, thus the check is technically incorrect.

Fixes: 205021945
Fixes: 205972662
Test: manual
Change-Id: I8372988baeca1e2165ac5bd85359b413f57478d8